requirements.txt
black==22.6.0
click==8.1.3
Django==4.0.6
djangorestframework==3.13.1
isort==5.10.1
typing_extensions==4.3.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
decorator==5.1.1
dj-database-url==0.5.0
django-on-heroku==1.1.2
drf-spectacular==0.22.1
executing==0.8.3
inflection==0.5.1
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
jsonschema==4.6.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
uritemplate==4.1.1
wcwidth==0.2.5
whitenoise==6.2.0